Move TagRequests into /domain/tags/model.py
This continues the DDD refactor, moving the other model and removing the top level models.py.
